+++ test/rocprim/test_device_batch_memcpy.cpp
@@ -326,12 +326,15 @@ TYPED_TEST(RocprimDeviceBatchMemcpyTests, SizeAndTypeVariation)
         common::device_ptr<value_type*>      d_buffer_dsts(num_buffers);
         common::device_ptr<buffer_size_type> d_buffer_sizes(num_buffers);
 
+        // Test if batch_copy also supports ptr-to-const type output.
+        const auto dsts = d_buffer_dsts.get();
+
         // Calculate temporary storage
         size_t temp_storage_bytes = 0;
         batch_copy<isMemCpy>(nullptr,
                              temp_storage_bytes,
                              d_buffer_srcs.get(),
-                             d_buffer_dsts.get(),
+                             dsts,
                              d_buffer_sizes.get(),
                              num_buffers,
                              hipStreamDefault,