+++ test/rocprim/test_transform_iterator.cpp
@@ -35,6 +35,7 @@
 #include <rocprim/device/device_reduce.hpp>
 #include <rocprim/functional.hpp>
 #include <rocprim/iterator/transform_iterator.hpp>
+#include <rocprim/iterator/transform_output_iterator.hpp>
 
 #include <cstddef>
 #include <numeric>
@@ -61,15 +62,11 @@ struct plus_ten
 };
 
 // Params for tests
-template<
-    class InputType,
-    class UnaryFunction = times_two<InputType>,
-    class ValueType = InputType
->
+template<class InputType, class UnaryFunction = times_two<InputType>, class ValueType = InputType>
 struct RocprimTransformIteratorParams
 {
-    using input_type = InputType;
-    using value_type = ValueType;
+    using input_type     = InputType;
+    using value_type     = ValueType;
     using unary_function = UnaryFunction;
 };
 
@@ -77,9 +74,9 @@ template<class Params>
 class RocprimTransformIteratorTests : public ::testing::Test
 {
 public:
-    using input_type = typename Params::input_type;
-    using value_type = typename Params::value_type;
-    using unary_function = typename Params::unary_function;
+    using input_type             = typename Params::input_type;
+    using value_type             = typename Params::value_type;
+    using unary_function         = typename Params::unary_function;
     const bool debug_synchronous = false;
 };
 
@@ -230,18 +227,144 @@ TYPED_TEST(RocprimTransformIteratorTests, Basic)
     }
 }
 
+TYPED_TEST(RocprimTransformIteratorTests, BasicOutput)
+{
+    int device_id = test_common_utils::obtain_device_from_ctest();
+    SCOPED_TRACE(testing::Message() << "with device_id = " << device_id);
+    HIP_CHECK(hipSetDevice(device_id));
+
+    // Use value type over input type due to precision/casting shenanigans.
+    using value_type     = typename TestFixture::value_type;
+    using unary_function = typename TestFixture::unary_function;
+    using iterator_type  = typename rocprim::transform_output_iterator<value_type*, unary_function>;
+    using difference_type = typename iterator_type::difference_type;
+
+    for(size_t seed_index = 0; seed_index < number_of_runs; seed_index++)
+    {
+        unsigned int seed_value
+            = seed_index < random_seeds_count ? rand() : seeds[seed_index - random_seeds_count];
+        SCOPED_TRACE(testing::Message() << "with seed = " << seed_value);
+
+        std::vector<value_type> input
+            = test_utils::get_random_data_wrapped<value_type>(10, 1, 200, seed_value);
+
+        std::vector<value_type> data = std::vector<value_type>(10, value_type{});
+
+        unary_function transform;
+
+        auto output
+            = rocprim::make_transform_output_iterator<value_type*, unary_function>(data.data(),
+                                                                                   transform);
+
+        auto output_begin = output;
+        auto output_mid   = output_begin + 5;
+        auto output_end   = output_begin + 10;
+
+        auto result_begin = data.begin();
+        auto result_end   = result_begin + 10;
+
+        // Pre-increment
+        auto it_o = output_begin;
+        auto it_r = result_begin;
+        ++it_o;
+        ++it_r;
+        *it_o = input[1];
+        ASSERT_EQ(*it_r, transform(input[1]));
+
+        // Post-increment
+        auto post_o = it_o++;
+        auto post_r = it_r++;
+        *it_o       = input[2];
+        ASSERT_EQ(*post_r, transform(input[1]));
+        ASSERT_EQ(*it_r, transform(input[2]));
+
+        // Pre-decrement
+        it_o = it_o + 2;
+        it_r = it_r + 2;
+        --it_o;
+        --it_r;
+        *it_o = input[3];
+        ASSERT_EQ(*it_r, transform(input[3]));
+
+        // Post-decrement
+        post_o = it_o--;
+        post_r = it_r--;
+        *it_o  = input[2];
+        ASSERT_EQ(*post_r, transform(input[3]));
+        ASSERT_EQ(*it_r, transform(input[2]));
+
+        // Pre-decrement via -=
+        it_o -= 2;
+        it_r -= 2;
+        *it_o = input[0];
+        ASSERT_EQ(*it_r, transform(input[0]));
+
+        // operator+
+        auto plus_it_o = output_begin + 3;
+        auto plus_it_r = result_begin + 3;
+        *plus_it_o     = input[3];
+        ASSERT_EQ(*plus_it_r, transform(input[3]));
+
+        auto plus_it_rev_o = 3 + output_begin;
+        auto plus_it_rev_r = 3 + result_begin;
+        *plus_it_rev_o     = input[3];
+        ASSERT_EQ(*plus_it_rev_r, transform(input[3]));
+
+        // operator-
+        auto minus_it_o = output_end - 3;
+        auto minus_it_r = result_end - 3;
+        *minus_it_o     = input[7];
+        ASSERT_EQ(*minus_it_r, transform(input[7]));
+
+        // compound assignment +=
+        auto a_o = output_begin;
+        auto a_r = result_begin;
+        a_o += 4;
+        a_r += 4;
+        *a_o = input[4];
+        ASSERT_EQ(*a_r, transform(input[4]));
+
+        // compound assignment -=
+        a_r -= 2;
+        a_o -= 2;
+        *a_o = input[2];
+        ASSERT_EQ(*a_r, transform(input[2]));
+
+        // Subtraction of iterators (distance)
+        ASSERT_EQ(output_end - output_begin, difference_type(10));
+        ASSERT_EQ(output_mid - output_begin, difference_type(5));
+        ASSERT_EQ(output_begin - output_mid, difference_type(-5));
+
+        // Indexing operator[]
+        for(int i = 0; i < 10; i++)
+        {
+            output_begin[i] = input[i];
+            ASSERT_EQ(result_begin[i], transform(input[i]));
+        }
+
+        // Comparisons
+        ASSERT_TRUE(output_begin == output_begin);
+        ASSERT_TRUE(output_begin != output_end);
+        ASSERT_TRUE(output_begin < output_end);
+        ASSERT_TRUE(output_end > output_begin);
+        ASSERT_TRUE(output_begin <= output_begin);
+        ASSERT_TRUE(output_begin <= output_end);
+        ASSERT_TRUE(output_end >= output_begin);
+        ASSERT_TRUE(output_end >= output_end);
+    }
+}
+
 TYPED_TEST(RocprimTransformIteratorTests, TransformReduce)
 {
     int device_id = test_common_utils::obtain_device_from_ctest();
     SCOPED_TRACE(testing::Message() << "with device_id = " << device_id);
     HIP_CHECK(hipSetDevice(device_id));
 
-    using input_type = typename TestFixture::input_type;
-    using value_type = typename TestFixture::value_type;
+    using input_type     = typename TestFixture::input_type;
+    using value_type     = typename TestFixture::value_type;
     using unary_function = typename TestFixture::unary_function;
-    using iterator_type = typename rocprim::transform_iterator<
-        input_type*, unary_function, value_type
-    >;
+    using iterator_type =
+        typename rocprim::transform_iterator<input_type*, unary_function, value_type>;
 
     hipStream_t stream = 0; // default
 
@@ -249,7 +372,8 @@ TYPED_TEST(RocprimTransformIteratorTests, TransformReduce)
 
     for(size_t seed_index = 0; seed_index < number_of_runs; seed_index++)
     {
-        unsigned int seed_value = seed_index < random_seeds_count  ? rand() : seeds[seed_index - random_seeds_count];
+        unsigned int seed_value
+            = seed_index < random_seeds_count ? rand() : seeds[seed_index - random_seeds_count];
         SCOPED_TRACE(testing::Message() << "with seed = " << seed_value);
 
         // Generate data
@@ -260,12 +384,12 @@ TYPED_TEST(RocprimTransformIteratorTests, TransformReduce)
         common::device_ptr<input_type> d_input(input);
         common::device_ptr<value_type> d_output(output.size());
 
-        auto reduce_op = rocprim::plus<value_type>();
+        auto           reduce_op = rocprim::plus<value_type>();
         unary_function transform;
 
         // Calculate expected results on host
         iterator_type x(input.data(), transform);
-        value_type expected = std::accumulate(x, x + size, value_type(0), reduce_op);
+        value_type    expected = std::accumulate(x, x + size, value_type(0), reduce_op);
 
         auto d_iter = iterator_type(d_input.get(), transform);
 
@@ -288,5 +412,4 @@ TYPED_TEST(RocprimTransformIteratorTests, TransformReduce)
         // Check if output values are as expected
         test_utils::assert_near(output[0], expected, test_utils::precision<value_type> * size);
     }
-
 }